Refactor simulation.py for more flexibility
This adds several features to the Simulation class.  First, it differentiates between running and executing, which means that processing can be done without halting a workflow (making SUMMA run in the background).  Second it stores standard and error outputs in properties, making it easier to debug SUMMA runs. It also has a status property for quickly checking the state of the object.  Finally, this provides cleaner interoperability with the configuration refactor, making it easier to get output files.